+++ run_unit_tests.sh
@@ -1,3 +1,3 @@
-python setup.py bdist_wheel
-pip install dist/*.whl
+#python setup.py bdist_wheel
+#pip install dist/*.whl
 python -m unittest discover -s test -p '*Test.py'
