--- test/test_logger.py
@@ -1,17 +1,0 @@
-import unittest
-import logging
-import sys
-
-root = logging.getLogger()
-root.setLevel(logging.DEBUG)
-
-handler = logging.StreamHandler(sys.stdout)
-handler.setLevel(logging.DEBUG)
-formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
-handler.setFormatter(formatter)
-root.addHandler(handler)
-
-
-class Test(unittest.TestCase):
-    def setUp(self):
-        logging.info('setup logger')
